distro/rhel9: add rhel-9 alias to rhel-90
Add a plain `rhel-9` alias as the default distribution name and version for the
`rhel9` package.  The `rhel-90` distro is still available via the
NewRHEL90() constructor.  These two distributions are identical.

Repositories
------------

The rhel-90 repositories are copied to rhel-9.json.  The files are
identical and point to the CDN repositories with no minor version (plain
`9` repositories).