weldr: read the host distro from the registry's FromHost()
The registry already reads the host distro information and sets the
value.  This can be retrieved with the registry's FromHost() method, so
we don't need to read the info again.

If the host distro isn't supported, or the host information couldn't be
read, print an appropriate message in the log.

The only functional change in this commit is that osbuild-composer can
still run the weldr API after failing to read the host distro
information.