dnf-json: explicitly set priority
Prioritise the repositories in the order they are passed. Currently their priorities are all
the same, which gives (to me) unexpected results.

The logic is now that appending a repo to the list of repositories, means that it overrides
the previously passed ones. This is relevant in case the same packages exist in the
base repositories as in the payload repositories. In this case we should prefer the
payload ones.