test/diff-manifests: fix error handling
The change made in 7f563a6db1a4fec4a00f2ddce8d9827b4abda6ab would
require the shell option `-e` to not be set, so that we could capture
the exit code after the command fails.
Fix the error handling by putting the commands that we want to handle in
the test part of an `if` clause.

In addition, error messages are now printed in red.