test/oscap.sh: provision SUT for the on-premise scenario
The PR#2695 [1] was not rebased on top of PR#2853 [2]. As a result, the
`oscap.sh` test case is provisioning the SUT in a wrong way, which makes
it fail all the time.

Provision the SUT in the test case for the on-premise scenario, since it
uses Weldr API to interact with osbuild-composer.

[1] https://github.com/osbuild/osbuild-composer/pull/2695
[2] https://github.com/osbuild/osbuild-composer/pull/2853