cloudapi: make share_with_accounts optional
With the introduction of `public`, it doesn't make sense to require
`share_with_accounts`, because it doesn't make sense to share a public AMI.
Therefore, this commit no longer marks share_with_accounts as required.

In order to keep things sane, the API handler now checks whether either
share_with_accounts is set and non-empty, or public is set to true. If this
isn't true, the request is rejected.

Unit tests are added in order to check this.

Signed-off-by: Ondřej Budai <ondrej@budai.cz>